lavf/qsv: release the frame in filter_frame().
We must release the input frame in filter_frame().
but for filters like overlay, which is based on framesync.
the input frame was managed by framesync.c.
To handle this problem, we clone the frame in qsvvpp.c
when filtering a frame. So as we are adding an frame_clone()
in qsvvpp.c, we need to add back the frame release
for filters that is not based on framesync.

Signed-off-by: Ruiling Song <ruiling.song@intel.com>